# apps/web/apps/web-antd/src/api/generated/infra.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RustSet Infra API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/infra/service-provider/page": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "pageNo",
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "default": null
            },
            "style": "form"
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "default": null
            },
            "style": "form"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_for_Page_for_AnyValue"
                }
              }
            }
          }
        }
      }
    },
...
    "/infra/service-provider/create": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceProviderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_for_string"
                }
              }
            }
          }
        }
      }
    },
    "/infra/service-provider/update": {
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServiceProviderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_for_null"
                }
              }
            }
          }
        }
      }
    },
...
    "/infra/machine-room/page": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "pageNo",
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "default": null
            },
            "style": "form"
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "default": null
            },
            "style": "form"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_for_Page_for_AnyValue"
                }
              }
            }
          }
        }
      }
    },
...
    "/infra/machine-room/create": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMachineRoomRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_for_string"
                }
              }
            }
          }
        }
      }
    },
    "/infra/machine-room/update": {
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMachineRoomRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_for_null"
                }
              }
            }
          }
        }
      }
    },
...
    "/infra/security-product/page": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "pageNo",
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "default": null
            },
            "style": "form"
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "default": null
            },
            "style": "form"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_for_Page_for_AnyValue"
                }
              }
            }
          }
        }
      }
    },
...
    "/infra/security-product/create": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSecurityProductRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_for_string"
                }
              }
            }
          }
        }
      }
    },
    "/infra/security-product/update": {
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSecurityProductRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_for_null"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_for_Page_for_AnyValue": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "uint16",
            "maximum": 65535,
            "minimum": 0
          },
          "data": {
            "$ref": "#/components/schemas/Page_for_AnyValue"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "data",
          "message"
        ]
      },
...
      "ApiResponse_for_null": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "uint16",
            "maximum": 65535,
            "minimum": 0
          },
          "data": {
            "type": "null"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "data",
          "message"
        ]
      },
      "ApiResponse_for_string": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "uint16",
            "maximum": 65535,
            "minimum": 0
          },
          "data": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "data",
          "message"
        ]
      },
...
      "CreateMachineRoomRequest": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string"
          },
          "areaSize": {
            "type": [
              "string",
              "null"
            ]
          },
          "cabinetCount": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "contactPerson": {
            "type": "string"
          },
          "contactPhone": {
            "type": "string"
          },
          "facilityType": {
            "type": "string"
          },
          "floor": {
            "type": [
              "string",
              "null"
            ]
          },
          "providerId": {
            "type": "integer",
            "format": "int64"
          },
          "remarks": {
            "type": [
              "string",
              "null"
            ]
          },
          "roomCode": {
            "type": "string"
          },
          "roomName": {
            "type": "string"
          },
          "roomType": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "default": "active"
          }
        },
        "additionalProperties": false,
        "required": [
          "roomName",
          "roomCode",
          "facilityType",
          "address",
          "providerId",
          "roomType",
          "contactPerson",
          "contactPhone"
        ]
      },
...
      "CreateSecurityProductRequest": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string"
          },
          "cloudPlatformId": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "contactPerson": {
            "type": "string"
          },
          "contactPhone": {
            "type": "string"
          },
          "deploymentMode": {
            "type": "string"
          },
          "features": {
            "type": [
              "string",
              "null"
            ]
          },
          "licenseExpiry": {
            "type": [
              "string",
              "null"
            ]
          },
          "licenseType": {
            "type": "string"
          },
          "machineRoomId": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "managementIp": {
            "type": [
              "string",
              "null"
            ]
          },
          "model": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "providerId": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "remarks": {
            "type": [
              "string",
              "null"
            ]
          },
          "serialNumber": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string",
            "default": "active"
          },
          "throughput": {
            "type": [
              "string",
              "null"
            ]
          },
          "vendor": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        },
        "additionalProperties": false,
        "required": [
          "name",
          "category",
          "vendor",
          "model",
          "version",
          "licenseType",
          "deploymentMode",
          "contactPerson",
          "contactPhone"
        ]
      },
      "CreateServiceProviderRequest": {
        "type": "object",
        "properties": {
          "businessLicense": {
            "type": "string"
          },
          "contactEmail": {
            "type": "string"
          },
          "contactPerson": {
            "type": "string"
          },
          "contactPhone": {
            "type": "string"
          },
          "headquarters": {
            "type": "string"
          },
          "logoUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "providerCode": {
            "type": "string"
          },
          "providerName": {
            "type": "string"
          },
          "remarks": {
            "type": [
              "string",
              "null"
            ]
          },
          "serviceArea": {
            "type": "string"
          },
          "shortName": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "default": "active"
          }
        },
        "additionalProperties": false,
        "required": [
          "providerName",
          "providerCode",
          "shortName",
          "contactPerson",
          "contactPhone",
          "contactEmail",
          "headquarters",
          "serviceArea",
          "businessLicense"
        ]
      },
...
      "Page_for_AnyValue": {
        "type": "object",
        "properties": {
          "list": {
            "type": "array",
            "items": true
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "list",
          "total"
        ]
      },
...
      "UpdateMachineRoomRequest": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string"
          },
          "areaSize": {
            "type": [
              "string",
              "null"
            ]
          },
          "cabinetCount": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "contactPerson": {
            "type": "string"
          },
          "contactPhone": {
            "type": "string"
          },
          "facilityType": {
            "type": "string"
          },
          "floor": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "providerId": {
            "type": "integer",
            "format": "int64"
          },
          "remarks": {
            "type": [
              "string",
              "null"
            ]
          },
          "roomCode": {
            "type": "string"
          },
          "roomName": {
            "type": "string"
          },
          "roomType": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "default": "active"
          }
        },
        "additionalProperties": false,
        "required": [
          "id",
          "roomName",
          "roomCode",
          "facilityType",
          "address",
          "providerId",
          "roomType",
          "contactPerson",
          "contactPhone"
        ]
      },
...
      "UpdateSecurityProductRequest": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string"
          },
          "cloudPlatformId": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "contactPerson": {
            "type": "string"
          },
          "contactPhone": {
            "type": "string"
          },
          "deploymentMode": {
            "type": "string"
          },
          "features": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "licenseExpiry": {
            "type": [
              "string",
              "null"
            ]
          },
          "licenseType": {
            "type": "string"
          },
          "machineRoomId": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "managementIp": {
            "type": [
              "string",
              "null"
            ]
          },
          "model": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "providerId": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "remarks": {
            "type": [
              "string",
              "null"
            ]
          },
          "serialNumber": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string",
            "default": "active"
          },
          "throughput": {
            "type": [
              "string",
              "null"
            ]
          },
          "vendor": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        },
        "additionalProperties": false,
        "required": [
          "id",
          "name",
          "category",
          "vendor",
          "model",
          "version",
          "licenseType",
          "deploymentMode",
          "contactPerson",
          "contactPhone"
        ]
      },
      "UpdateServiceProviderRequest": {
        "type": "object",
        "properties": {
          "businessLicense": {
            "type": "string"
          },
          "contactEmail": {
            "type": "string"
          },
          "contactPerson": {
            "type": "string"
          },
          "contactPhone": {
            "type": "string"
          },
          "headquarters": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "logoUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "providerCode": {
            "type": "string"
          },
          "providerName": {
            "type": "string"
          },
          "remarks": {
            "type": [
              "string",
              "null"
            ]
          },
          "serviceArea": {
            "type": "string"
          },
          "shortName": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "default": "active"
          }
        },
        "additionalProperties": false,
        "required": [
          "id",
          "providerName",
          "providerCode",
          "shortName",
          "contactPerson",
          "contactPhone",
          "contactEmail",
          "headquarters",
          "serviceArea",
          "businessLicense"
        ]
      }
    }
  }
}